establish project structure
This is the initial commit for the motd tool.
It establishes the project structure, splitting things
up between a library crate and a binary that wraps
it in a CLI tool. This is just a system component
of shpool, so it is already approved to be open-sourced.
I'm putting it in a seperate repo since it could in
theory be used by someone else for something unrelated.