use walkdir to find pam_motd.so (#2)
This patch switches to using the walkdir crate to find pam_motd.so,
and changes the API so that we cache the result. Taken together
these changes go a long way to mitigating the fact that we need
to root around in /usr/lib to find pam_motd.so. walkdir is much
faster than the naive code that I wrote as part of my first
wack at things and caching the result means that a user does not
need to pay for the walk each time (they technically could have
cached with the old API, but it would have been pretty gross,
better to do it for them).

Also, I'm going to go ahead and configure CI while I'm at it.