add pure rust implementation
This patch adds a pure rust alternative to calling
the .so file. It turns out that our internal google
toolchain injects a custom linker which makes the
.so file calling shennigans unsafe due to potential
libc version skew. Having a pure rust version that
avoids the shennigans at the cost of potential
divergence from the system pam_motd.so allows us
to work around the issue.